requirements.txt
flask>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=1.9.0
PyPDF2>=2.0.0
python-docx>=0.8.11
transformers>=4.30.0
torch>=2.0.0
accelerate